+++ CHANGELOG.md
@@ -115,7 +115,11 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - Added Integration Tests for the entire IRS flow using stubbed responses of Discovery Service, Semantic Hub, EDC,
   Digital Twin Registry and BPDM Pool. #344
 
-### Changed
+## Added
+
+- Cucumber test step definitions for Policy Store API (Happy Path) including some test helper utilities. #518
+
+## Changed
 
 - Dataspace Discovery Service handles multiple EDC-Urls received for BPN now. #214
 - Updated license header to "Copyright (c) 2021,2024 Contributors to the Eclipse Foundation" #349